[Breaking] Generator.prototype
This proposes adding a JS class called `Generator` which is the type of all generators created in this library.

This also adds a handful of methods to `Generator.prototype` to provide a more idomatic API for working with Generators:

* `generator.then()` is a replacement for `gen.map()` and `gen.bind()` - it works similarly to Promise#then in that it can return either another Generator or any value.

* `generate.where()` is a replacement for `gen.suchThat()` - it accepts a predicate.

* `generate.sample()` is a replacement for `sample()`.

In addition to this, this PR creates an internal wrapping and unwrapping API which converts non-generators into generators. This allows for `gen.return()` to be rarely used, reduces the duplication of functions, and allows for tuples or objects to use both plain values and generators at the same time.