Fix uniqueArray() unique on nullish values
If the generator provided to `gen.uniqueArray()` could produce `undefined`, `null`, or `NaN` then these values could appear more than once. Also, arrays and objects of values are not considered for uniqueness.

This introduces a helper function `js-key-fn` which ensures values generated can be determined to be distinct for CLJ's Set.

Fixes #60