s/Generator/ValueGenerator/g
Using the name `Generator` is problematic because of its ambiguity with the JS Generator instances (the result of calling a generator function). This can lead to accidental type errors in typescript and flow when using the `Generator` type and forgetting to import it.

To correct for this, this diff introduces `ValueGenerator` as a new name for this class. The existing export continues to exist with a deprecation flag so as not to break existing test code, however the types have been updated so strict typed code will need to be updated.