Print seed in integration results
Following up on #66 - this prints the seed used when encountering failure to other test integrations, and unifies some behavior in printing between them.

Printing the seed during failure is useful when attempting to reproduce a failing test case on a rare case, or just repeating the same test to consistently reproduce a specific condition. Each integration offers a way to use the printed seed as the specific seed used in a test case.

Fixes #43